1. Gamepads support was improved; 2. Playground was integrated with gamepads; 3. Tweaks in actors turning for gamepads; 4. Helpers for deadzone handling.

+++ levels/gravity/gravity.cpp
@@ -241,7 +241,7 @@ namespace Levels
 
 			player1Controls.turningDelta = mouse.getWorldSpaceDelta();
 			player1Controls.autoRotation = mouse.pressing.rmb;
-			player1Controls.throttling = mouse.pressing.rmb;
+			player1Controls.throttling = (float)mouse.pressing.rmb;
 			player1Controls.magneticHook = mouse.pressing.xmb1;
 
 			if (mouse.pressing.lmb)
